refactor: upgrade fetch implementation to utilize enhanced route handler in Next.js v13
