+++ app/layout.tsx
@@ -2,12 +2,8 @@
 
 import { CssBaseline, createTheme } from "@mui/material";
 import StyledComponentsRegistry from "lib/registry";
-import { Inter } from "next/font/google";
-import Head from "next/head";
 import { ThemeProvider } from "styled-components";
 import "./globals.css";
-
-const inter = Inter({ subsets: ["latin"] });
 
 export default function RootLayout({
   children,
@@ -18,14 +14,10 @@ export default function RootLayout({
 
   return (
     <html lang="en">
-      <Head>
-        <meta name="viewport" content="initial-scale=1, width=device-width" />
-      </Head>
-
       <StyledComponentsRegistry>
         <ThemeProvider theme={theme}>
           <CssBaseline />
-          <body className={inter.className}>{children}</body>
+          <body>{children}</body>
         </ThemeProvider>
       </StyledComponentsRegistry>
     </html>
